feat: seperated the subgroup components out into files
